First Step into Responsive
The original Nike.com site isn't very responsive. I want to improve on that. So first, I had to take their banner images (Which was 2 images, a top and a bottom which split one of the pictures in half) Combine that into one image, then split it into 4 images, one for each picture. Then it was a matter of styling to get the images in the right place, (an offset 4x4 grid) then have them split into a single column once you go below 800px in width.